Fix for thimble.mozilla.org#2237
Fixed mozilla/thimble.mozilla.org issue #2237 - After deleting a file, it's name is still indicated in the Editor header bar.
Earlier, when last file was deleted, it was returning from the function without calling on the editor change event.
Right now, I have added a workaround for this, by sending an object with filename as "No File" which gets set in editor-nav-pane.